Split up api/plotting.ts into figure.ts and io.ts
